Fixed a bug in ScrollPane.ScrollToView Fixed a bug in MovieClip

+++ FairyGUI-egret/src/Transition.ts
@@ -826,14 +826,14 @@ module fairygui {
                         value.b1 = false;
                     }
                     else {
-                        value.f1 = parseInt(arr[0]);
+                        value.f1 = parseFloat(arr[0]);
                         value.b1 = true;
                     }
                     if(arr[1] == "-") {
                         value.b2 = false;
                     }
                     else {
-                        value.f2 = parseInt(arr[1]);
+                        value.f2 = parseFloat(arr[1]);
                         value.b2 = true;
                     }
                     break;
